getDocumentFromComponent Fixes
This update fixes `null` handing for `getDocumentFromComponent`, specifically
for React v16.

The solution was to improve `null` handling for `dash-get`, the lower-level
`get` module used to retrieve the desired `document` object.

`dash-get` has been updated to `v1.0.2` with the fix and additional tests
have been added to check for `null` within the internal React instance
object.